requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.2
flask-sqlalchemy==3.1.1
Flask-JWT-Extended==4.5.3
eventlet==0.35.2
opencv-python==4.8.1.78
mediapipe==0.10.21
psycopg2-binary==2.9.9 
python-dotenv==1.0.1
numpy==1.26.3
pandas==2.2.3